refactor(api): migrate correction serializer to evaluation context
