+++ api/src/evaluation/application/answers/answer-controller.js
@@ -1,8 +1,8 @@
 import { usecases } from '../../../../lib/domain/usecases/index.js';
-import * as correctionSerializer from '../../../../lib/infrastructure/serializers/jsonapi/correction-serializer.js';
 import * as requestResponseUtils from '../../../../lib/infrastructure/utils/request-response-utils.js';
 import { evaluationUsecases } from '../../domain/usecases/index.js';
 import * as answerSerializer from '../../infrastructure/serializers/jsonapi/answer-serializer.js';
+import * as correctionSerializer from '../../infrastructure/serializers/jsonapi/correction-serializer.js';
 
 const save = async function (request, h, dependencies = { answerSerializer, requestResponseUtils }) {
   const answer = dependencies.answerSerializer.deserialize(request.payload);
